Implement message modal with OAM.Modal
Remove the now obsolete base modal
